chore: Position, Address에 대해 common에서 별도의 디렉토리로 분리
