+++ src/main/java/kr/pickple/back/address/domain/AddressDepth2.java
@@ -1,4 +1,4 @@
-package kr.pickple.back.common.domain;
+package kr.pickple.back.address.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,6 +6,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotNull;
+import kr.pickple.back.common.domain.BaseEntity;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
