+++ src/main/java/kr/pickple/back/game/domain/Game.java
@@ -16,8 +16,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
-import kr.pickple.back.common.domain.AddressDepth1;
-import kr.pickple.back.common.domain.AddressDepth2;
+import kr.pickple.back.address.domain.AddressDepth1;
+import kr.pickple.back.address.domain.AddressDepth2;
 import kr.pickple.back.common.domain.BaseEntity;
 import kr.pickple.back.member.domain.Member;
 import lombok.AccessLevel;
