+++ src/main/java/kr/pickple/back/position/domain/Position.java
@@ -1,4 +1,4 @@
-package kr.pickple.back.common.domain;
+package kr.pickple.back.position.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,12 +6,13 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotNull;
+import kr.pickple.back.common.domain.BaseEntity;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AddressDepth1 extends BaseEntity {
+public class Position extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -20,4 +21,12 @@ public class AddressDepth1 extends BaseEntity {
     @NotNull
     @Column(unique = true, length = 10)
     private String name;
+
+    @NotNull
+    @Column(length = 100)
+    private String description;
+
+    @NotNull
+    @Column(unique = true, length = 2)
+    private String acronym;
 }
